PreTrainedModel/Whisper/flask_api/requirements.txt
flask[async]
gunicorn
torch
openai-whisper
python-multipart
pydub